% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiGradient.R
\name{niftiGradient}
\alias{niftiGradient}
\title{NIfTI Gradient}
\usage{
niftiGradient(image, which = "all")
}
\arguments{
\item{image}{an image of class \code{\link{nifti}}}

\item{which}{a string specifying the gradient direction that should be returned. Either "all" for a list of x, y, and z gradients, or "x", "y", or "z" for a single image with the given gradient.}
}
\value{
Either a list of three gradient images or a single gradient image.
}
\description{
This function returns the gradient images for a 3-dimensional NIfTI image.
}
\examples{
\dontrun{
library(neurobase)
epi <- readnii('path/to/epi')
gradients <- niftiGradient(image = epi, which = "all") }
}
