% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiHessian.R
\name{niftiHessian}
\alias{niftiHessian}
\title{NIfTI Hessian}
\usage{
niftiHessian(image, mask = NULL, parallel = FALSE)
}
\arguments{
\item{image}{an image of class \code{\link{nifti}}}

\item{mask}{a mask of class \code{\link{nifti}},
if \code{NULL} the hessian filter will be run for the full array.
Note that mask should be in the same space as the image volume}

\item{parallel}{is a logical value that indicates whether the user's computer
is Linux or Unix (i.e. macOS), and should run the code in parallel}
}
\value{
A list of three eigenvalue volumes.
}
\description{
This function returns the eigenvalues of the hessian matrices for a 3-dimensional NIfTI volume.
}
\examples{
\dontrun{
library(neurobase)
epi <- readnii('path/to/epi')
mask <- epi!=0
hesseigs <- niftiHessian(image = epi, mask = mask) }
}
